+++ src/popup/Header/Sidebar.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import './styles.scss'
-
-const Sidebar = () => {
-  return <div className='settings__page'>settings</div>
-}
-
-export default Sidebar