+++ CHANGELOG.md
@@ -2,6 +2,4 @@
 
 ## src/popup
 
-- add tooltip for the header buttons
-- added drawer to show the settins
-- add the option to chanage the header color
+- add select field on top of rule title and regex check box